refactor: split OBR APIs/background jobs
- Break the monolithic OBR integration modules into focused API and background task files to improve maintainability and make scheduler wiring clearer. Update all client-side method calls and hook scheduler paths to match the new module layout, and clean up formatting across touched Python/JS files.

- Key behavior updates include using item descriptions in invoice and stock movement payload item designations (falling back to code/batch when missing) and fixing Sales Invoice custom field visibility condition syntax (`frm.doc` -> `doc`) for dependency evaluation.
