+++ burundi_compliance/burundi_compliance/apis/stock_movement.py
@@ -1,0 +1,8 @@
+import frappe
+
+from ..background_tasks.tasks import send_stock_movement_to_obr
+
+
+@frappe.whitelist()
+def trigger_stock_movement_to_obr() -> None:
+	send_stock_movement_to_obr()
