+++ burundi_compliance/burundi_compliance/utils/build_invoice_payload.py
@@ -183,7 +183,9 @@ def get_invoice_items(doc):
 				total_vat += tax.get("amount", 0)
 
 		item_designation = (
-			f"{item.item_code}-{item.batch_no}" if item.batch_no else item.item_code
+			item.description
+			if item.description
+			else (f"{item.item_code}-{item.batch_no}" if item.batch_no else item.item_code)
 		)
 
 		items.append(
@@ -202,62 +204,3 @@ def get_invoice_items(doc):
 		)
 
 	return items
-
-	# for item in doc.items:
-	#     tax_data = next(
-	#         (data for data in itemised_tax_data if data["item"] == item.item_code),
-	#         None,
-	#     )
-	#     total_vat = 0
-
-	#     if tax_data:
-	#         # Check if VAT exists, if not, check for other tax details
-	#         if "VAT" in tax_data:
-	#             total_vat = tax_data["VAT"]["tax_amount"]
-	#         else:
-	#             total_vat = 0
-	#     else:
-	#         total_vat = 0
-	# item_designation = (
-	#     item.item_code + "-" + item.batch_no if item.batch_no else item.item_code
-	# )
-	# items.append(
-	#     {
-	#         "item_code": item.item_code,
-	#         "item_designation": item_designation,
-	#         "item_quantity": abs(item.qty),
-	#         "item_price": item.rate,
-	#         "item_total_amount": item.amount,
-	#         "vat": abs(int(total_vat)),
-	#         "item_ct": "0",
-	#         "item_tl": "0",
-	#         "item_price_nvat": abs(int(item.amount)),
-	#         "item_price_wvat": abs(int(item.amount + total_vat)),
-	#     }
-	# )
-	# return items
-
-
-# [
-#     {
-#         "account_head": "VAT - MS",
-#         "name": "de76jio985",
-#         "creation": datetime.datetime(2026, 4, 10, 13, 28, 58, 327608),
-#         "modified": datetime.datetime(2026, 4, 10, 13, 28, 58, 327608),
-#         "modified_by": "Administrator",
-#         "owner": "Administrator",
-#         "docstatus": 1,
-#         "idx": 1,
-#         "item_row": "ckl72qon67",
-#         "tax_row": "d9see3b9dc",
-#         "rate": 18.0,
-#         "amount": 180.0,
-#         "taxable_amount": 1000.0,
-#         "parent": "ABC-2026-04-10-00004",
-#         "parentfield": "item_wise_tax_details",
-#         "parenttype": "Sales Invoice",
-#     }
-# ]
-
-
-# wsl400038739100326
