+++ pyproject.toml
@@ -1,7 +1,7 @@
 [project]
 name = "burundi_compliance"
 authors = [
-    { name = "Navari Limited", email = "mania@navari.co.ke"}
+    { name = "Navari Limited", email = "solutions@navari.co.ke"}
 ]
 description = "Burundi OBR Tax Integration"
 requires-python = ">=3.10"
@@ -11,7 +11,7 @@ dependencies = [
     #"frappe~=15.0.0", # Installed and managed by bench.
     "qrcode~=7.4.2",
     "PyQRCode==1.2.1",
-    "pillow==10.2.0",
+    "pillow==11.0.0",
     "pypng==0.20220715.0"
 ]
